refactor: extract sibling modules from performance, detection, error_handling
- performance_cache.py: CacheEntry/CacheStats/Cache/RequestOptimizer;
  performance.py keeps ConnectionPool (ServerProxy patch target),
  PerformanceMonitor, PerformanceManager + compat re-exports
- detection_probes.py: ProbeResult, _timed, the 8 probe functions;
  detection.py keeps _PROBES (patch target), _aggregate, entry points
- error_handler.py: ErrorHandler + handle_odoo_error/format_user_error;
  error_handling.py keeps the MCPError hierarchy and re-exports lazily
  via module __getattr__ to avoid an import cycle

All files ≤450 lines; existing import paths preserved.
